Reorganizing Standalone runner and RestIntegTest task
