+++ buildSrc/src/main/java/org/elasticsearch/gradle/test/RestIntegTestTask.java
@@ -19,101 +19,13 @@
 
 package org.elasticsearch.gradle.test;
 
-import org.elasticsearch.gradle.testclusters.ElasticsearchCluster;
-import org.elasticsearch.gradle.testclusters.TestClustersAware;
-import org.elasticsearch.gradle.testclusters.TestClustersThrottle;
-import org.elasticsearch.gradle.util.GradleUtils;
-import org.gradle.api.Task;
-import org.gradle.api.provider.Provider;
-import org.gradle.api.services.internal.BuildServiceRegistryInternal;
+import org.elasticsearch.gradle.testclusters.StandaloneRestIntegTestTask;
 import org.gradle.api.tasks.CacheableTask;
-import org.gradle.api.tasks.Internal;
-import org.gradle.api.tasks.Nested;
-import org.gradle.api.tasks.testing.Test;
-import org.gradle.internal.resources.ResourceLock;
-import org.gradle.internal.resources.SharedResource;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.List;
-
-import static org.elasticsearch.gradle.testclusters.TestClustersPlugin.THROTTLE_SERVICE_NAME;
 
 /**
- * Customized version of Gradle {@link Test} task which tracks a collection of {@link ElasticsearchCluster} as a task input. We must do this
- * as a custom task type because the current {@link org.gradle.api.tasks.TaskInputs} runtime API does not have a way to register
- * {@link Nested} inputs.
+ * Sub typed version of {@link StandaloneRestIntegTestTask}  that is used to differentiate between plain standalone
+ * integ test tasks based on {@link StandaloneRestIntegTestTask} and
+ * conventional configured tasks of {@link RestIntegTestTask}
  */
 @CacheableTask
-public class RestIntegTestTask extends Test implements TestClustersAware {
-
-    private Collection<ElasticsearchCluster> clusters = new HashSet<>();
-
-    public RestIntegTestTask() {
-        this.getOutputs()
-            .doNotCacheIf(
-                "Caching disabled for this task since it uses a cluster shared by other tasks",
-                /*
-                 * Look for any other tasks which use the same cluster as this task. Since tests often have side
-                 * effects for the cluster they execute against, this state can cause issues when trying to cache
-                 * tests results of tasks that share a cluster. To  avoid any undesired behavior we simply disable
-                 * the cache if we detect that this task uses a cluster shared between multiple tasks.
-                 */
-                t -> getProject().getTasks()
-                    .withType(RestIntegTestTask.class)
-                    .stream()
-                    .filter(task -> task != this)
-                    .anyMatch(task -> Collections.disjoint(task.getClusters(), getClusters()) == false)
-            );
-    }
-
-    @Override
-    public int getMaxParallelForks() {
-        return 1;
-    }
-
-    @Nested
-    @Override
-    public Collection<ElasticsearchCluster> getClusters() {
-        return clusters;
-    }
-
-    @Override
-    @Internal
-    public List<ResourceLock> getSharedResources() {
-        List<ResourceLock> locks = new ArrayList<>(super.getSharedResources());
-        BuildServiceRegistryInternal serviceRegistry = getServices().get(BuildServiceRegistryInternal.class);
-        Provider<TestClustersThrottle> throttleProvider = GradleUtils.getBuildService(serviceRegistry, THROTTLE_SERVICE_NAME);
-        SharedResource resource = serviceRegistry.forService(throttleProvider);
-
-        int nodeCount = clusters.stream().mapToInt(cluster -> cluster.getNodes().size()).sum();
-        if (nodeCount > 0) {
-            locks.add(resource.getResourceLock(Math.min(nodeCount, resource.getMaxUsages())));
-        }
-
-        return Collections.unmodifiableList(locks);
-    }
-
-    @Override
-    public Task dependsOn(Object... dependencies) {
-        super.dependsOn(dependencies);
-        for (Object dependency : dependencies) {
-            if (dependency instanceof Fixture) {
-                finalizedBy(((Fixture) dependency).getStopTask());
-            }
-        }
-        return this;
-    }
-
-    @Override
-    public void setDependsOn(Iterable<?> dependencies) {
-        super.setDependsOn(dependencies);
-        for (Object dependency : dependencies) {
-            if (dependency instanceof Fixture) {
-                finalizedBy(((Fixture) dependency).getStopTask());
-            }
-        }
-    }
-}
+public class RestIntegTestTask extends StandaloneRestIntegTestTask {}
